GPIO refactoring in progress (#156)
* gpio module refactored

#153

* adding test cases fixing typos

* gpio module refactor

adding edgpi gpio test

* fix suggestions
